Reorganize data models into different folder
