+++ src/app/data-models/handle-model.ts
