+++ src/app/data-models/interval-model.ts
