+++ src/app/data-models/temperament-model.ts
