+++ src/app/data-models/vector-model.ts
